+++ CHANGELOG.rst
@@ -3,8 +3,9 @@ Changelog
 =========
 
 
-Unreleased
-----------
+v0.3.1
+------
+*04 February 2021*
 
 Fixed
 ~~~~~